+++ includes/figures.h
@@ -14,8 +14,8 @@ typedef struct s_sphere
 {
 	t_vector_4	vectors[MAX_VECTORS_SPHERE];
 	t_color		color;
-	float		radius_squared;
-	float		specular;
+	double		radius_squared;
+	double		specular;
 	t_byte		traits;
 }				t_sphere;
 
@@ -24,7 +24,7 @@ typedef struct s_plane
 	t_vector_4	position;
 	t_vector_4	normal;
 	t_color		color;
-	float		specular;
+	double		specular;
 	t_byte		traits;
 }				t_plane;
 
@@ -38,8 +38,8 @@ typedef struct s_cylinder
 {
 	t_vector_4	vectors[MAX_VECTORS_CYLINDER];
 	t_color		color;
-	float		radius2;
-	float		specular;
+	double		radius2;
+	double		specular;
 	t_byte		traits;
 }				t_cylinder;
 
@@ -53,10 +53,10 @@ typedef struct s_cone
 {
 	t_vector_4	vectors[MAX_VECTORS_CONE];
 	t_color		color;
-	float		angle;
-	float		cos;
-	float		cos2;
-	float		specular;
+	double		angle;
+	double		cos;
+	double		cos2;
+	double		specular;
 	t_byte		traits;
 }				t_cone;
