+++ includes/raytrace.h
@@ -16,7 +16,7 @@ typedef struct s_pre_light
 	t_vector_4		P;
 	t_vector_4		N;
 	t_vector_4		V;
-	float			intensity;
+	double			intensity;
 }				t_pre_light;
 
 void	process_pixel(t_rtv *rtv, short xc, short yc);
